chore(style): run nix fmt

+++ overlays/20-recurse-overrides.nix
@@ -1,20 +1,34 @@
-self: super:
-{
-  dix = super.recurseIntoAttrs (super.dix or { });
+self: super: {
+  dix = super.recurseIntoAttrs (super.dix or {});
 
-  zed-editor = if super.stdenv.isDarwin then null else super.recurseIntoAttrs super.zed-editor;
+  zed-editor =
+    if super.stdenv.isDarwin
+    then null
+    else super.recurseIntoAttrs super.zed-editor;
 
-  cudaPackages = super.recurseIntoAttrs (super.cudaPackages // {
-    cudnn = if super.stdenv.isDarwin then null else
-    (
-      super.cudaPackages.cudnn.overrideAttrs (oa: {
-        postFixup = oa.postFixup + ''
-          rm $out/LICENSE
-        '';
-      })
-    );
-    tensorrt = if super.stdenv.isDarwin then null else super.cudaPackages.tensorrt;
-  });
+  cudaPackages = super.recurseIntoAttrs (super.cudaPackages
+    // {
+      cudnn =
+        if super.stdenv.isDarwin
+        then null
+        else
+          (
+            super.cudaPackages.cudnn.overrideAttrs (oa: {
+              postFixup =
+                oa.postFixup
+                + ''
+                  rm $out/LICENSE
+                '';
+            })
+          );
+      tensorrt =
+        if super.stdenv.isDarwin
+        then null
+        else super.cudaPackages.tensorrt;
+    });
 
-  cudatoolkit = if super.stdenv.isDarwin then null else super.recurseIntoAttrs super.cudatoolkit;
+  cudatoolkit =
+    if super.stdenv.isDarwin
+    then null
+    else super.recurseIntoAttrs super.cudatoolkit;
 }
